Changed the attributes of functions in the LLVM IR
Every function is declared as 'nounwind' because the LLVM backend is
supposed to use setjmp/longjmp based exception handling.

Additionally, to make the generated native assembly easier to read, the
functions are declared as 'noinline', so LLVM does not inline any calls.
This may be reverted when the main debugging phase is over.