The generated LLVM IR now contains calls to gcroot
For that to work I added code to work around what seems to be a bug in
LLVM occuring when using gcroot with the ocaml gc.

Also a bug concerning exception handling that caused an infinite loop in
certain when an exception was thrown in a try-with-block that was not
caught in that block.

Furthermore changed the runtime code a bit to work with the new
exception handling.