Removed lots of code from the old native code backend that the new one does not need
This includes options for dumping intermediate data as well es the data
structures used to represent that data.  In particular Mach and
Linearize are replaced by a variante used for generating LLVM IR.  Any
code used for register allocation as well as a lot of amd64-specific
code are removed completely.