Some changes for debugging the interface to the garbage collector
The type inference on the Cmm code now assumes every let-binding can
contain a pointer.  Every argument of function calls and alloc
instructions is computed, saved on the stack.  They are loaded as soon
as they are needed to call the function/store the values.  It seemed
like this might fix the GC-problem, but it did not.  During garbage
collection some strange bugs occur...